tweak: swap to async-safe `HashMap`/`HashSet` from `scc`
- swap last_pint to use an atomic cell instead of a RwLock to reduce any potential locking
- swap udp_socket_addr to use an ArcSwapOption to reduce locking
- swap `write` to use a Mutex and swapped it to an async safe lock
- swap `crypt_state` to use Mutex instead of a RwLock since it would already only be used as a write lock
- reduced some reference counting

+++ src/server/constants.rs
@@ -1,5 +1,5 @@
 /// The amount of players the server can support
-pub const MAX_CLIENTS: u32 = 4096;
+pub const MAX_CLIENTS: usize = 4096;
 
 /// the bandwidth (in bytes) that the client can use
 pub const MAX_BANDWIDTH: u32 = 144000;
